+++ include/sdk/interfaces/common/CThreadSpinRWLock.h
@@ -6,34 +6,20 @@
 #include "sdk/interfaces/common/CThreadMutex.h"
 #include <type_traits>
 
-#if defined(CS2) || defined(DOTA2)
-constexpr auto kThreadSpinMutex = 2;
-#else
-constexpr auto kThreadSpinMutex = 1;
-#endif
-
-class CThreadSpinMutexV1 {
+class CThreadSpinRWLock {
 public:
-    explicit CThreadSpinMutexV1(const char* pDebugName = nullptr): m_ownerID(0), m_depth(0), m_pszDebugName(pDebugName) { }
+    struct LockInfo_t {
+        std::uint32_t m_writerId;
+        std::int32_t m_nReaders;
+    };
 
-private:
-    volatile ThreadId_t m_ownerID;
-    int m_depth;
+public:
+    void* m_pThreadSpin;
+    LockInfo_t m_lockInfo;
     const char* m_pszDebugName;
 };
 
-class CThreadSpinMutexV2 {
-public:
-    explicit CThreadSpinMutexV2([[maybe_unused]] const char* pDebugName = nullptr): m_ownerID(0), m_depth(0) { }
-
-private:
-    volatile ThreadId_t m_ownerID;
-    int m_depth;
-};
-
-using CThreadSpinMutex = std::conditional_t<kThreadSpinMutex == 1, CThreadSpinMutexV1, CThreadSpinMutexV2>;
-using CThreadFastMutex = CThreadSpinMutex;
-
+static_assert(sizeof(CThreadSpinRWLock) == 0x18);
 // source2gen - Source2 games SDK generator
 // Copyright 2024 neverlosecc
 //
