+++ include/sdk/interfaces/common/CUtlMemory.h
@@ -143,6 +143,8 @@ public:
     int m_nGrowSize;
 };
 
+static_assert(sizeof(CUtlMemory<int, int>) == 0x10);
+
 //-----------------------------------------------------------------------------
 // constructor, destructor
 //-----------------------------------------------------------------------------
